Code Updates after First Submission Feedback
1. Removed the marked out module variable call at the beginning and end
of the file. This should now declare the entire game as “ticTacToe”

2. Comments now included in the global variable section and reorganized
to represent how the variables are placed in the start, board, and
finish elements of the game.

3. Rearranged the file to better define areas of functions, code to run
on load, and event listeners. Sub categories are also created within
the “Declared Function” section which help to better describe what the
section of functions mainly do.

4. Moved some of the original code within the “clickBoxes” function
into its new separate function entitled ‘afterClickCheck’.

5. Created the function playerGameCheck which replaces the original
creation of having a separate “check” function for each player. This
will now run for both players with the symbol lookup changing
dynamically based on the current click count in the game.

6. Created the ‘gameReset’ function which completely resets the game
technically and visually to its original starting point. This function
can now be used in other ways such as if we wanted to create a “Reset
Game” button within the game board, we could utilize and call this
function to make the reset happen.

7. General comment updates